// apps/web/src/generated/types/Course.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const courseSemesterEnum = {
    "fall": "fall",
    "spring": "spring"
} as const;

export type CourseSemesterEnum = (typeof courseSemesterEnum)[keyof typeof courseSemesterEnum];

export type Course = {
    /**
     * @type string
    */
    id: string;
    /**
     * @type string
    */
    code: string;
    /**
     * @type string
    */
    name: string;
    /**
     * @type string
    */
    description: string;
    /**
     * @type integer
    */
    credits: number;
    /**
     * @type string
    */
    instructorId: string;
    /**
     * @type string
    */
    semester: CourseSemesterEnum;
    /**
     * @type integer
    */
    year: number;
    /**
     * @type array
    */
    tags: string[];
    /**
     * @type integer
    */
    capacity: number;
    /**
     * @type integer
    */
    enrolled: number;
};